Extract Sentry initialization
To resolve a circular dependency (reportException uses the store
to provide username etc. to Sentry, so importing it from within
the store was causing trouble), created src/initSentry, which
abstracts the initialization logic so that components within
the store can report errors to Sentry without causing a circular
import.
